+++ crates/menmos-repository/src/lib.rs
@@ -3,7 +3,7 @@ mod iface;
 mod s3;
 pub mod util;
 
-pub use iface::Repository;
+pub use iface::{OperationGuard, Repository};
 
 pub use disk::DiskRepository;
 pub use s3::S3Repository;
